refactor: Initialize shift with number and date generators
