+++ spec/enigma_spec.rb
@@ -9,17 +9,7 @@ RSpec.describe Enigma do
     expect(@enigma).to be_a (Enigma)
   end
 
-  it 'creates an array of 4 key strings from the key' do
-    expect(@enigma.key_to_four_pairs.length).to eq (4)
-    expect(@enigma.key_to_four_pairs).to be_a (Array)
-    four_keys = []
-    four_keys << @enigma.key[0..1].to_i
-    four_keys << @enigma.key[1..2].to_i
-    four_keys << @enigma.key[2..3].to_i
-    four_keys << @enigma.key[3..4].to_i
-    expect(@enigma.key_to_four_pairs).to eq four_keys
-    expect(@enigma.key_to_four_pairs).to eq []
-  end
+  
 
   it 'creates an array of 4 offsets based on the date' do
     expect(@enigma.convert_offset("040895")).to be_a (Array)
